requirements.txt
flask
sqlalchemy
psycopg2-binary
alembic
pyjwt